Refactor: move `authorize.py` and `auth.py`
Mostly changing imports.
